fixed validation

+++ proj_man_sys/labor_costs/views.py
@@ -8,34 +8,46 @@ from django.contrib.auth.models import User
 
 
 def create_labor_costs(request):
-    if request.method == "POST":
-        form = CreateLaborForm(request.POST or None, request.FILES or None)
-        if form.is_valid():
-            form.save()
-            labor = Labor.objects.latest("id")
-            labor.executor = request.user.last_name + " " + request.user.first_name
-            labor.save()
-            return redirect("view_labor_costs")
-    if request.method == "GET":
-        form = CreateLaborForm()
-        form.fields["task"].queryset = Task.objects.filter(users=request.user)
-        return render(request, "add_or_change_labor.html", {"form": form})
+    if not request.user.is_anonymous:
+        if request.method == "POST":
+            form = CreateLaborForm(request.POST or None, request.FILES or None, request = request)
+            if form.is_valid():
+                form.save()
+                labor = Labor.objects.latest("id")
+                labor.executor = request.user.last_name + " " + request.user.first_name
+                labor.save()
+                return redirect("view_labor_costs")
+            else:
+                error_message = 'Некорректные данные'
+                form = CreateLaborForm(request = request)
+                return render(request, "add_or_change_labor.html", {"form": form, "error_message":error_message })
+        else:
+            form = CreateLaborForm(request = request)
+            return render(request, "add_or_change_labor.html", {"form": form})
+    else:
+        return redirect('add_or_change_proj')
 
 
 def view_labor_costs(request):
-    labor_costs = Labor.objects.all()
-    admin = request.user.last_name + " " + request.user.first_name
-    return render(
-        request,
-        "labor_costs_view.html",
-        {"labor_costs": labor_costs, "admin": admin},
-    )
+    if not request.user.is_anonymous:
+        labor_costs = Labor.objects.all()
+        admin = request.user.last_name + " " + request.user.first_name
+        return render(
+            request,
+            "labor_costs_view.html",
+            {"labor_costs": labor_costs, "admin": admin},
+        )
+    else:
+        return redirect('add_or_change_proj')
 
 
 def delete_labor(request, id):
-    try:
-        labor = Labor.objects.get(id=id)
-        labor.delete()
-        return redirect("view_labor_costs")
-    except Labor.DoesNotExist:
-        return HttpResponseNotFound("<h1>Ресурс не найден</h1>")
+    if not request.user.is_anonymous:
+        try:
+            labor = Labor.objects.get(id=id)
+            labor.delete()
+            return redirect("view_labor_costs")
+        except Labor.DoesNotExist:
+            return HttpResponseNotFound("<h1>Ресурс не найден</h1>")
+    else:
+        return redirect('add_or_change_proj')
